pyproject.toml
[project]
name = "slack-dashboard"
version = "0.2.0"
description = "Slack thread monitoring dashboard with heat-ranked conversations and LLM summarization"
readme = "README.md"
authors = [
    { name = "Scott A. Idler", email = "scott.a.idler@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "anthropic>=0.86.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "slack-sdk>=3.41.0",
    "uvicorn>=0.42.0",
]

[project.scripts]
slack-dashboard = "slack_dashboard.main:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
    "types-markdown>=3.10.2.20260211",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
